Configure renovate to upgrade crossplane components
Renovate has built-in support for kustomize helm chart inflation but
only when using the `helmCharts` property.

A CODEOWNERS file is introduced so the renovate PRs will be auto-assigned
when opened.

The version specific info in the README is removed along with the
unneccessary note about loading images for kind. The latter is only
required when using local images which isn't applicable.

The dependency dashboard is also disabled because it's unused.

Signed-off-by: Alex Misstear <amisstea@redhat.com>